added in ghsci.py, which addresses #276, previously unintentionally omitted, but perhaps ideally, over-writing the older file on which it was based _project_setup.py; this one allows construction of Regions as an object containing configuration settings
